Moved ObjectOIDEscapement out of the ObjectOID folder. The ObjectOID folder will only be used for implementations of ObjectOID from now on.
