+++ AtelierTomato.Markov/AtelierTomato.Markov.Data/Model/ObjectOIDEscapement.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.RegularExpressions;
 
-namespace AtelierTomato.Markov.Data.Model.ObjectOID
+namespace AtelierTomato.Markov.Data.Model
 {
 	public static class ObjectOIDEscapement
 	{
